requirements.txt
twitchio
tinydb
python-dotenv